--- a/db/migrate/20200428143250_add_primary_key_to_rules_references.rb
+++ b/db/migrate/20200428143250_add_primary_key_to_rules_references.rb
@@ -1,0 +1,5 @@
+class AddPrimaryKeyToRulesReferences < ActiveRecord::Migration[5.2]
+  def change
+    add_column :rule_references_rules, :id, :primary_key
+  end
+end